requirements.txt
# requirements.txt
Flask==3.0.0
flask-cors==4.0.0
openai-whisper==20231117
python-docx==1.1.0
fpdf==1.7.2
moviepy==1.0.3
werkzeug==3.0.1
gunicorn==21.2.0
torch==2.1.0
numpy==1.24.3
ffmpeg-python==0.2.0
